Updating agent code and file name to match course
